refactor(desktop): extract connections ipc
